+++ CHANGELOG.rst
@@ -10,6 +10,12 @@ Unreleased__
 ------------
 __ https://github.com/israel-lugo/capidup/compare/v1.0.1...HEAD
 
+Added
+.....
+
+- `find_duplicates_in_dirs` can now exclude directories, through a new optional
+  parameter `exclude_dirs`. See `issue #10`_.
+
 Changed
 .......
 
@@ -31,6 +37,7 @@ First production release.
 
 
 .. _issue #9: https://github.com/israel-lugo/capidup/issues/9
+.. _issue #10: https://github.com/israel-lugo/capidup/issues/10
 .. _issue #12: https://github.com/israel-lugo/capidup/issues/12
 
 .. _1.0.1: https://github.com/israel-lugo/capidup/tree/v1.0.1